% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSBP.R
\name{control_Gibbs}
\alias{control_Gibbs}
\title{Control parameters for the Gibbs sampling algorithm}
\usage{
control_Gibbs(R = 5000, burn_in = 1000, method_init = "cluster")
}
\arguments{
\item{R}{An integer indicating the number of MCMC iterations to be computed after the \code{burn-in}.}

\item{burn_in}{An integer indicating the number of MCMC iterations discarded as burn-in period.}

\item{method_init}{The initialization method. The default \code{method_init='cluster'} partitions the covariates  using the \code{\link[cluster]{clara}} clustering algorithm.  Other available options are: \code{method_init='random'} and \code{method_init='deterministic'}.}
}
\value{
The inputs are converted into a \code{list}. Missing arguments are filled with default values.
}
\description{
The \code{control_Gibbs} function can be used for specifying the technical settings (i.e. the number of MCMC iterations, the burn-in, and the initialization method), of the \code{\link[LSBP]{LSBP_Gibbs}} function.
}
