% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DLSBP.R
\name{control_VB}
\alias{control_VB}
\title{Control parameters for the VB algorithm}
\usage{
control_VB(maxiter = 10000, tol = 0.01, method_init = "cluster")
}
\arguments{
\item{maxiter}{An integer indicating the maximum number of iterations for the ECM algorithm.}

\item{tol}{A real number controlling the convergence of the algorithm. It is defined as the (absolute) difference of the log-posterior between consecutive iterations that has to be reached.}

\item{method_init}{The initialization criterium. By default, \verb{method_init='cluster'} preallocates covariates into groups using \code{\link[cluster]{clara}}. Another available possibility is \verb{method_init='random'}.}
}
\value{
The function returns a list having the same entries provided as argument. Missing arguments are filled with default values.
}
\description{
Control parameters for the VB algorithm
}
