% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DLSBP.R
\name{DLSBP_ECM}
\alias{DLSBP_ECM}
\title{ECM algorithm for the DLSBP model}
\usage{
DLSBP_ECM(Formula, data = NULL, H, prior, control = control_ECM(),
  verbose = TRUE)
}
\arguments{
\item{Formula}{An object of class \code{\link[Formula]{Formula}}: a symbolic description of the model to be fitted. The details of model specification are given under 'Details'.}

\item{data}{A data frame containing the variables of \verb{Formula}.}

\item{H}{An integer indicating the number of mixture components.}

\item{prior}{A list of prior hyperparameters as returned by \code{\link[DLSBP]{prior_DLSBP}}. If missing, default prior values are used.}

\item{control}{A list as returned by \code{\link[DLSBP]{control_ECM}}.}

\item{verbose}{A logical value indicating whether additional information should be displayed while the algorithm is running.}
}
\value{
The output is an object of class '\verb{DLSBP_ECM}' containing the following quantities:
\itemize{
\item \verb{param}. A list containing the MAP (Maximum A Posteriori), for each set of coefficients: \verb{beta_mixing,beta_kernel,tau}.
\item \verb{cluster}. A n dimensional vector containing, for each observation, the mixture component having with the highest probability.
\item \verb{z}. A \verb{n x H} matrix containing the probabilities of belonging to each of the mixture components, where \verb{n} denotes the number of observations.
\item \verb{logposterior}. The log-posterior of the model at convergence.
\item \verb{call}. The input Formula 
\item \verb{data}. The input data frame.
\item \verb{control}. The control list provided as input.
\item \verb{H}. The input number of mixture components.
\item \verb{prior}. The input prior hyperparameters.
}
}
\description{
The dependent logit stick-breaking process (DLSBP) model is estimated using the E(C)M algorithm, which provides the posterior mode.
}
\details{
The \verb{Formula} specification contains the response, separated from the covariates with the symbol '\verb{~}', and two sets of covariates. The latters are separated by the symbol '\verb{|}', indicating the kernel covariates and the mixing covariates, respectively. For example, one could specify \verb{y ~ x1 + x2 | x3 + x4}. NOTE: if the second set of covariates is omitted it is assumed that the two sets are the same.

If \verb{offsets} or \verb{weights} are provided in the \verb{Formula} they will be ignored in the current version.
A \verb{predict} method is available and described at \code{\link[DLSBP]{predict.DLSBP_ECM}}.
}
\examples{
data(cars)

# A model with constant kernels
fit_em <- DLSBP_ECM(dist ~  1 | speed, data=cars, H=4)
plot(cars) 
lines(cars$speed,predict(fit_em))

# A model with linear kernels
fit_em <- DLSBP_ECM(dist ~ speed | speed, data=cars, H=2)
plot(cars) 
lines(cars$speed,predict(fit_em))

}
\references{
Rigon, T. and Durante, D., (2017), Logit stick-breaking priors for Bayesian density regression, ArXiv.
}
