% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSBP.R
\name{LSBP_Pois_Gibbs}
\alias{LSBP_Pois_Gibbs}
\title{Gibbs sampling for the LSBP model with Poisson Kernels}
\usage{
LSBP_Pois_Gibbs(Formula, data, H, prior, control = control_Gibbs(),
  verbose = TRUE)
}
\description{
Gibbs sampling for the LSBP model with Poisson Kernels
}
