% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSBP.R
\name{LSBP_Pois_VB}
\alias{LSBP_Pois_VB}
\title{Variational Bayes algorithm for the LSBP model with Poisson Kernels}
\usage{
LSBP_Pois_VB(Formula, data, H, prior, control = control_VB(),
  verbose = TRUE)
}
\description{
Variational Bayes algorithm for the LSBP model with Poisson Kernels
}
