% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{predict.DLSBP_VB}
\alias{predict.DLSBP_VB}
\title{Predict method for the DLSBP estimated using VB}
\usage{
\method{predict}{DLSBP_VB}(object, type = "mean", R = 5000,
  newdata = NULL, ...)
}
\description{
Predict method for the DLSBP estimated using VB
}
