% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{predict.LSBP_Gibbs}
\alias{predict.LSBP_Gibbs}
\title{Predict method for the LSBP}
\usage{
\method{predict}{LSBP_Gibbs}(object, type = "mean", newdata = NULL,
  threshold = NULL, ...)
}
\arguments{
\item{object}{An object of class \verb{LSBP_Gibbs}.}

\item{type}{String indicating the type of prediction: \verb{type="mean"},\verb{type="predictive"},\verb{type="variance"} or \verb{type="cdf"}. See details.}

\item{newdata}{A new data frame containing the same variables declared in \verb{Formula}. If missing, the dataset provided for estimation is used.}

\item{threshold}{Only needed if \verb{type="cdf"} is selected. See details.}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Predict method for the LSBP estimated using the Gibbs sampling algorithm.
}
\details{
The method \verb{predict.LSBP_Gibbs} produces a sample of predicted values, obtained by evaluating the conditional mean of the LSBP model or the predictive distribution, using the observations contained in the \verb{newdata} data frame. 

If \verb{type="mean"} a sample from the posterior distribution of the LSBP mean is returned. If \verb{type="predictive"} is selected, then a sample from the predictive distribution is returned.  If \verb{type="variance"} a sample from the posterior distribution of the LSBP variance is returned.  If \verb{type="cdf"} a sample from the posterior distribution of the LSBP cumulative distribution function is returned, evaluated at \verb{threshold}.
}
