% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSBP.R
\name{control_ECM}
\alias{control_ECM}
\title{Control parameters for the ECM algorithm}
\usage{
control_ECM(maxiter = 10000, tol = 0.001, method_init = "cluster")
}
\arguments{
\item{maxiter}{An integer indicating the maximum number of iterations for the \code{\link[LSBP]{LSBP_ECM}} algorithm.}

\item{tol}{A real number controlling the convergence of the algorithm. The \code{\link[LSBP]{LSBP_ECM}} algorithm stops when the difference between consecutive values of the log-posterior is smaller than \code{tol}.}

\item{method_init}{The initialization method. The default \code{method_init='cluster'} partitions the covariates  using the \code{\link[cluster]{clara}} clustering algorithm. Other available options are: \code{method_init='random'} and \code{method_init='deterministic'}.}
}
\value{
The inputs are converted into a \code{list}. Missing arguments are filled with default values.
}
\description{
The \code{control_ECM} function can be used for specifying the technical settings (i.e. the maximum number of iterations, the tolerance level, and the initialization method), of the \code{\link[LSBP]{LSBP_ECM}} function.
}
