% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{DLSBP_EM}
\alias{DLSBP_EM}
\title{Bayesian nonparametric density estimation}
\usage{
DLSBP_EM(y, X, H = 2, maxiter = 1000, prior = NULL, verbose = TRUE,
  tol = 1e-04, random = FALSE, pstep = 10)
}
\arguments{
\item{y}{A vector containing the response vector}

\item{X}{A n \times p design matrix containing the covariates}

\item{H}{The number of mixtures}

\item{maxiter}{Logical: should be putted a prior on alpha?}

\item{prior}{Logical: should be putted a prior on alpha?}

\item{verbose}{Logical: should be putted a prior on alpha?}

\item{tol}{Logical: should be putted a prior on alpha?}

\item{random}{Logical: should be putted a prior on alpha?}

\item{pstep}{Logical: should be putted a prior on alpha?}
}
\description{
The dependent logistic stick-breaking process (DLSBP) model posterior mod, estimated trhough the EM algorithm
}

