% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{DLSBP_Gibbs}
\alias{DLSBP_Gibbs}
\title{Bayesian nonparametric density estimation}
\usage{
DLSBP_Gibbs(y, X, H = 2, R = 5000, burn.in = 2000, prior = NULL,
  verbose = TRUE, parameters = TRUE, useEM = TRUE, pstep = 100)
}
\arguments{
\item{y}{A vector containing the response vector}

\item{X}{A n \times p design matrix containing the covariates}

\item{H}{The number of mixtures}

\item{prior}{Logical: should be putted a prior on alpha?}

\item{verbose}{Logical: should be putted a prior on alpha?}

\item{pstep}{Logical: should be putted a prior on alpha?}

\item{maxiter}{Logical: should be putted a prior on alpha?}

\item{tol}{Logical: should be putted a prior on alpha?}

\item{random}{Logical: should be putted a prior on alpha?}
}
\description{
The dependent logistic stick-breaking process (DLSBP) model posterior mod, estimated trhough the EM algorithm
}

