% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{LSBP_density}
\alias{LSBP_density}
\title{Conditional density of a LSBP model}
\usage{
LSBP_density(y, X1, X2, beta_mixing, beta_kernel, tau)
}
\arguments{
\item{y}{The value at which the conditional density must be evaluated.}

\item{X1}{A \code{n x p_kernel} design matrix for the kernel.}

\item{X2}{A \code{n x p_mixing} design matrix for the stick-breaking weights.}

\item{beta_mixing}{A \code{H-1 x p_mixing} dimensional matrix of coefficients for the linear predictor of the stick-breaking weights.}

\item{beta_kernel}{A \code{H x p_kernel} dimensional matrix of coefficients for the linear predictor of the kernel.}

\item{tau}{A \code{H} dimensional vector of coefficients for the kernel precision.}
}
\description{
Evaluate the conditional density \eqn{f_x(y)} of a LSBP model, given the parameters and the covariates.
}
\details{
The function \code{LSBP_density} evaluates the conditional density \eqn{f_x(y)}. The number of mixture components \code{H} is inferred from the dimensions of \code{beta_mixing} and \code{beta_kernel}.
}
