% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{LSBP_density}
\alias{LSBP_density}
\title{Predict method for the LSBP}
\usage{
LSBP_density(y, X1, X2, beta, alpha, tau)
}
\arguments{
\item{y}{A value of which the conditional density has to be computed}

\item{X1}{A \verb{n x p_kernel} design matrix for the kernel}

\item{X2}{A \verb{n x p_mixing} design matrix for the stick-breaking weights}

\item{beta}{A \verb{H x p_kernel} dimensional matrix of coefficients for the linear predictor of the kernel}

\item{alpha}{A \verb{H-1 x p_mixing} dimensional matrix of coefficients for the linear predictor of the stick-breaking weights}

\item{tau}{A \verb{H} dimensional vector of coefficients for the kernel precision}
}
\description{
Predict method for the LSBP estimated using the Variational Bayes algorithm.
}
\details{
The function \verb{LSBP_density} evaluates the conditional density of \verb{y} given the parameters
}
