% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSBP.R
\name{LSBP_Gibbs}
\alias{LSBP_Gibbs}
\title{Gibbs sampling algorithm for the LSBP model}
\usage{
LSBP_Gibbs(Formula, data, H, prior, control = control_Gibbs(),
  verbose = TRUE)
}
\arguments{
\item{Formula}{An object of class \code{\link[Formula]{Formula}}: a symbolic description of the model to be fitted. The details of model specification are given under 'Details'.}

\item{data}{A data frame containing the variables of \code{Formula}.}

\item{H}{An integer indicating the number of mixture components.}

\item{prior}{A list of prior hyperparameters as returned by \code{\link[LSBP]{prior_LSBP}}. If missing, default prior values are used.}

\item{control}{A list as returned by \code{\link[LSBP]{control_Gibbs}}.}

\item{verbose}{A logical value indicating whether additional information should be displayed while the algorithm is running.}
}
\value{
The output is an object of class '\code{LSBP_Gibbs}' containing the following quantities:
\itemize{
\item \code{param}. A list containing MCMC replications for each set of coefficients: \code{beta_mixing, beta_kernel, tau}.
\item \code{logposterior}. The \code{log-posterior} of the model at each MCMC iteration.
\item \code{call}. The input Formula.
\item \code{data}. The input data frame.
\item \code{control}. The control list provided as input.
\item \code{H}. The input number of mixture components.
\item \code{prior}. The input prior hyperparameters.
}
}
\description{
This function is an implementation of the Gibbs sampling Algorithm 1 in Rigon, T. and Durante, D. (2020).
}
\details{
The \code{Formula} specification contains the response, separated from the covariates with the symbol '\code{~}', and two sets of covariates. The latters are separated by the symbol '\code{|}', indicating the kernel covariates and the mixing covariates, respectively. For example, one could specify \code{y ~ x1 + x2 | x3 + x4}. NOTE: if the second set of covariates is omitted it is implicitely assumed that the two sets are the same.

If \code{offsets} or \code{weights} are provided in the \code{Formula} they will be ignored in the current version.
A \code{predict} method is available and described at \code{\link[LSBP]{predict.LSBP_Gibbs}}.
}
\examples{
\dontrun{
data(cars)

# A model with constant kernels
fit_gibbs <- LSBP_Gibbs(dist ~  1 | speed, data=cars, H=4)
plot(cars) 
lines(cars$speed,colMeans(predict(fit_gibbs)))

# A model with linear kernels
fit_gibbs <- LSBP_Gibbs(dist ~ speed | speed, data=cars, H=2)
plot(cars) 
lines(cars$speed,colMeans(predict(fit_gibbs)))
}

}
\references{
Rigon, T. and Durante, D., (2020), Tractable Bayesian density regression via logit stick-breaking priors. Journal of Statistical Planning and Inference.
}
