% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{predict.DLSBP_Gibbs}
\alias{predict.DLSBP_Gibbs}
\title{Predict method for the DLSBP}
\usage{
\method{predict}{DLSBP_Gibbs}(object, type = "mean", newdata = NULL, ...)
}
\arguments{
\item{object}{An object of class \verb{DLSBP_Gibbs}.}

\item{type}{Type of prediction that has to be computed. See details.}

\item{newdata}{A new data frame containing the same variables declared in \verb{Formula}. If missing, the dataset provided for estimation is used.}
}
\value{
The method \verb{predict.Gibbs} produces predicted values, obtained by evaluating the conditional mean of the DLSBP model or the predictive distribution, using the observations contained in the \verb{newdata} data frame.
}
\description{
Predict method for the DLSBP estimated using the Gibbs sampling algorithm.
}
\details{
Ciao
}
